fix(review): MEDIUMs finais (paginacao, pg_trgm, audit, bulk insert)
Cashier
- pendingReviews: paginacao (page/pageSize, default 0/20). Retorna
  { data, total, pageCount }. UI ajustada.
- openCashiers: limit defensivo de 200 + orderBy openedAt desc.
- forceClose: audit log explicito (action='force_close') registra
  calculatedBalance, original userId e reason. Rastreabilidade da
  acao que mascara divergencia.

Stock
- pg_trgm index em prod: name, sku, barcode, brand + customers.name.
  Acelera ILIKE/contains em tenants com 10k+ produtos.
- entrySerializedItems: trocou loop sequencial por createManyAndReturn
  + stockMovement.createMany. Ganho real em imports com 50+ aparelhos.

Validator
- csvImportLineSchema: costPrice/salePrice/promotionalPrice agora
  .int() (centavos). Bloqueia float que gerava Decimal fracionario.

+++ src/lib/validators/stock.ts
@@ -323,9 +323,11 @@ export const csvImportLineSchema = z.object({
   barcode: z.string().optional(),
   brand: z.string().optional(),
   category: z.string().optional(),
-  costPrice: z.number().min(0).optional(),
-  salePrice: z.number().min(0, "Preco de venda obrigatorio"),
-  promotionalPrice: z.number().min(0).optional(),
+  // Valores em CENTAVOS. int() bloqueia float que gera Decimal fracionario
+  // (ex: 1500.5 cents -> 15.005 reais -> arredonda inconsistente).
+  costPrice: z.number().int().min(0).optional(),
+  salePrice: z.number().int().min(0, "Preco de venda obrigatorio"),
+  promotionalPrice: z.number().int().min(0).optional(),
   minStock: z.number().int().min(0).optional(),
   quantity: z.number().int().min(0).optional(),
   isSerialized: z.boolean().optional(),
